+++ src/Adapter/User/UserAuthenticationMode.php
@@ -5,11 +5,14 @@ namespace FluxIliasRestApi\Adapter\User;
 enum UserAuthenticationMode: string
 {
 
+    case APACHE = "apache";
     case CAS = "cas";
     case DEFAULT = "default";
+    case ECS = "ecs";
     case LDAP = "ldap";
     case LOCAL = "local";
-    case OPENID = "openid";
+    case LTI = "lti";
+    case OPEN_ID_CONNECT = "open-id-connect";
     case RADIUS = "radius";
     case SAML = "saml";
     case SCRIPT = "script";